+++ src/ui/widgets/Dot.h
@@ -25,7 +25,7 @@ class Dot : public juce::Component, public HasStyle {
     }
 
     bool hasText() override { return false; }
-    void setTextSize(int) override {}
+    void setTextSize(double) override {}
     void setFont(juce::Font) override {}
     void setTextAlign(juce::Justification) override {}
