+++ src/ui/widgets/Slider.h
@@ -24,7 +24,7 @@ class Slider : public juce::Slider, public HasStyle {
     }
 
     bool hasText() override { return false; }
-    void setTextSize(int) override {}
+    void setTextSize(double) override {}
     void setFont(juce::Font) override {}
     void setTextAlign(juce::Justification) override {}
 };
